change lambda with broken typing to def
This function is annotated with a Callable[] syntax that works only
because of a bug with attribute methods in the current mypy. In a future
release of mypy this is fixed and this will start throwing an error. If
a lambda is needed the only way to type it correctly would be defining a
protocol and using __call__, but I've opted to replace the lambda with a
def instead since it's much simpler.